refactoring structure, fixed winner flex display
